% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_on_response.R
\name{derivatives}
\alias{derivatives}
\title{Compute the first and second order derivatives of the response function}
\usage{
derivatives(fit_obj, obs = NULL)
}
\arguments{
\item{fit_obj}{Fitted object}
}
\description{
Compute the first and second order derivatives of the response function
}
