% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poly.R
\name{fit_poly}
\alias{fit_poly}
\title{Polynomial regression}
\usage{
fit_poly(x, y, degree = 1)
}
\arguments{
\item{x}{Matrix of predictors}

\item{y}{Vector of response values}

\item{degree}{Degree of the polynomial}
}
\description{
Polynomial regression
}
