% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_on_response.R
\name{interactions}
\alias{interactions}
\title{Title}
\usage{
interactions(fit_obj, index_col1 = 1, index_col2 = 2, obs = NULL)
}
\arguments{
\item{fit_obj}{}

\item{index_col1}{}

\item{index_col2}{}

\item{obs}{}
}
\description{
Title
}
