% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.matern32}
\alias{plot.matern32}
\title{Plot method for matern32 objects}
\usage{
plot.matern32(
  fit_obj,
  choice = c("coeffs", "GCV", "heterogeneity", "interactions", "residuals"),
  ...
)
}
\arguments{
\item{fit_obj}{Fitted object of class 'matern32'}

\item{choice}{The type of plot to produce (coeffs, GCV, heterogeneity, interactions, residuals)}

\item{...}{}
}
\description{
Plot method for matern32 objects
}
