% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_matern32.R
\name{predict.matern32}
\alias{predict.matern32}
\title{Predict from Matern 3/2 model}
\usage{
\method{predict}{matern32}(
  fit_obj,
  newx,
  level = NULL,
  method = c("splitconformal", "kde", "surrogate", "bootstrap"),
  nsim = 250L,
  B = nsim
)
}
\arguments{
\item{fit_obj}{Fitted object from \code{fit_matern32}}

\item{newx}{New data matrix}

\item{level}{Confidence level for prediction intervals (between 0 and 100)}

\item{method}{Method to use for prediction intervals}

\item{nsim}{Number of simulations to use for prediction}

\item{B}{Alias for \code{nsim}}
}
\description{
Predict from Matern 3/2 model
}
\examples{

n <- 100 ; p <- 4

set.seed(456)
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

lams <- 10^seq(-10, 10, length.out = 50)

fit_obj <- fit_matern32(x = X, y = y, lambda = lams)

df <- data.frame(predict(fit_obj, X) - y)
colnames(df) <- paste0(round(lams, 2))
summary(df)
boxplot(df[, c(1, 10, 25, 35, 50)], 
main = "distribution of in sample bias", 
xlab = "lambda", ylab = "y_hat - y")

}
