% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poly.R
\name{predict.poly}
\alias{predict.poly}
\title{Predict method for polynomial regression}
\usage{
\method{predict}{poly}(object, newx)
}
\arguments{
\item{object}{Fitted object}

\item{newx}{New data}
}
\description{
Predict method for polynomial regression
}
