% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.matern32}
\alias{summary.matern32}
\title{Summary method for matern32}
\usage{
summary.matern32(fit_obj, obs = NULL, verbose = TRUE)
}
\arguments{
\item{fit_obj}{Fitted object of class 'matern32'}

\item{obs}{observation number for which to calculate the derivatives (default is NULL)}

\item{verbose}{logical indicating whether to print the summary (default is TRUE)}
}
\description{
Summary method for matern32
}
