% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_matern32.R
\name{fit_matern32}
\alias{fit_matern32}
\title{Fit Matern 3/2 model}
\usage{
fit_matern32(
  x,
  y,
  lambda = 10^seq(-10, 10, length.out = 100),
  l = NULL,
  method = c("chol", "solve", "svd", "eigen", "conformal"),
  with_kmeans = FALSE,
  centers = NULL,
  centering = FALSE,
  seed = 123,
  ...
)
}
\arguments{
\item{x}{Matrix of predictors}

\item{y}{Vector of response values}

\item{lambda}{Regularization parameter}

\item{l}{Hyperparameter for the Matern 3/2 kernel}

\item{method}{Method to use for fitting the model}

\item{with_kmeans}{Use k-means to reduce the number of observations}

\item{centers}{Number of centers to use in k-means}

\item{centering}{Center the response}

\item{seed}{Seed for reproducibility}

\item{...}{}

\item{sigma}{Hyperparameter for the Matern 3/2 kernel}
}
\description{
Fit Matern 3/2 model
}
\examples{

n <- 100; p <- 4

set.seed(456)
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

lams <- 10^seq(-5, 4, length.out = 50)

fit_obj <- matern32::fit_matern32(x = X, y = y, lambda = lams)

par(mfrow=c(1, 2))
 
plot(log(lams), fit_obj$GCV, type = 'l', main = "GCV", 
ylab = "GCV")

matplot(log(lams), t(fit_obj$coef), type = 'l', 
main = "coefficients = f(lambda)", xlab = "log(lambda)", 
ylab = "coefs")
abline(h = 0, lty = 2, lwd = 2, col = "red")

matern32::summary.matern32(fit_obj)


library(MASS)
X <-  longley[,-7]
y <- longley[, 7]
fit_obj <- matern32::fit_matern32(x = X, y = y)

par(mfrow=c(1, 2))
 
plot(log(fit_obj$lambda), fit_obj$GCV, type = 'l', main = "GCV", 
ylab = "GCV")

matplot(log(fit_obj$lambda), t(fit_obj$coef), type = 'l', 
main = "coefficients = f(lambda)", xlab = "log(lambda)", 
ylab = "coefs")
abline(h = 0, lty = 2, lwd = 2, col = "red")

matern32::summary.matern32(fit_obj)

library(MASS)

X <- as.matrix(Boston[,-14])
y <- Boston[,14]

fit_obj <- matern32::fit_matern32(x = X, y = y, lambda = lams)

matern32::summary.matern32(fit_obj)
 
}
