% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.matern32}
\alias{plot.matern32}
\title{Title}
\usage{
plot.matern32(
  fit_obj,
  choice = c("coeffs", "GCV", "heterogeneity", "interactions", "residuals"),
  ...
)
}
\arguments{
\item{fit_obj}{}

\item{choice}{}

\item{...}{}
}
\value{

}
\description{
Title
}
