% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Title}
\usage{
plot_interactions(
  fit_obj,
  var1 = 1,
  var2 = 2,
  method = c("mars", "poly"),
  degree = 1,
  ...
)
}
\arguments{
\item{fit_obj}{}

\item{var1}{}

\item{var2}{}

\item{...}{}
}
\description{
Title
}
