% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_matern32.R
\name{predict.matern32}
\alias{predict.matern32}
\title{Predict from Matern 3/2 model}
\usage{
\method{predict}{matern32}(fit_obj, newx, ci = NULL)
}
\arguments{
\item{fit_obj}{}

\item{newx}{}

\item{ci}{}
}
\description{
Predict from Matern 3/2 model
}
\examples{

n <- 100 ; p <- 4

set.seed(456)
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

lams <- 10^seq(-10, 10, length.out = 50)

fit_obj <- fit_matern32(x = X, y = y, lambda = lams)

df <- data.frame(predict(fit_obj, X) - y)
colnames(df) <- paste0(round(lams, 2))
summary(df)
boxplot(df[, c(1, 10, 25, 35, 50)], 
main = "distribution of in sample bias", 
xlab = "lambda", ylab = "y_hat - y")

}
