% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi.R
\name{sensi1D}
\alias{sensi1D}
\title{Title}
\usage{
sensi1D(fit_obj, index_col, h, verbose = TRUE)
}
\arguments{
\item{fit_obj}{object fitted by using \code{fit_matern32}}

\item{index_col}{an integer, column index}

\item{h}{(must be small)}
}
\value{

}
\description{
Title
}
