% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi.R
\name{sensi2D}
\alias{sensi2D}
\title{Title}
\usage{
sensi2D(fit_obj, index_col1, index_col2, h1, h2, verbose = TRUE)
}
\arguments{
\item{fit_obj}{object fitted by using \code{fit_matern32}}

\item{index_col1}{an integer, column index}

\item{index_col2}{an integer, second column index}

\item{h}{(must be small)}
}
\description{
Title
}
